requirements.txt
discord
flask
python-dotenv
waitress